+++ meridian/analysis/analyzer_test.py
@@ -788,6 +788,35 @@ class AnalyzerTest(tf.test.TestCase, parameterized.TestCase):
         atol=1e-3,
     )
 
+  def test_marginal_roi_media_and_rf_new_data_modified_times(self):
+    model.Meridian.inference_data = mock.PropertyMock(
+        return_value=self.inference_data_media_and_rf
+    )
+    mroi = self.analyzer_media_and_rf.marginal_roi(
+        new_data=analyzer.DataTensors(
+            media=self.meridian_media_and_rf.media_tensors.media[..., -15:, :],
+            media_spend=self.meridian_media_and_rf.media_tensors.media_spend[
+                ..., -15:, :
+            ],
+            reach=self.meridian_media_and_rf.rf_tensors.reach[..., -15:, :],
+            frequency=self.meridian_media_and_rf.rf_tensors.frequency[
+                ..., -15:, :
+            ],
+            rf_spend=self.meridian_media_and_rf.rf_tensors.rf_spend[
+                ..., -15:, :
+            ],
+            revenue_per_kpi=self.meridian_media_and_rf.revenue_per_kpi[
+                ..., -15:
+            ],
+        ),
+    )
+    self.assertAllClose(
+        mroi,
+        tf.convert_to_tensor(test_utils.MROI_MEDIA_AND_RF_NEW_TIMES_DATA),
+        rtol=1e-3,
+        atol=1e-3,
+    )
+
   def test_roi_wrong_new_data_dims_raises_exception(self):
     with self.assertRaisesWithLiteralMatch(
         ValueError,
@@ -813,8 +842,10 @@ class AnalyzerTest(tf.test.TestCase, parameterized.TestCase):
   def test_roi_wrong_new_data_times_raises_exception(self):
     with self.assertRaisesWithLiteralMatch(
         ValueError,
-        "New `rf_spend` is expected to have 49 time periods. Found 10 time"
-        " periods.",
+        "If the time dimension of a variable in `new_data` is modified, then"
+        " all variables must be provided in `new_data`. The following variables"
+        " are missing: `['media', 'media_spend', 'reach', 'frequency',"
+        " 'revenue_per_kpi']`.",
     ):
       self.analyzer_media_and_rf.roi(
           new_data=analyzer.DataTensors(
@@ -869,6 +900,35 @@ class AnalyzerTest(tf.test.TestCase, parameterized.TestCase):
         self.analyzer_media_and_rf.incremental_outcome() / total_spend
     )
     self.assertAllClose(expected_roi, roi)
+
+  def test_roi_media_and_rf_new_params_returns_correct_shape(self):
+    model.Meridian.inference_data = mock.PropertyMock(
+        return_value=self.inference_data_media_and_rf
+    )
+    roi = self.analyzer_media_and_rf.roi(
+        new_data=analyzer.DataTensors(
+            media=self.meridian_media_and_rf.media_tensors.media[..., -15:, :],
+            media_spend=self.meridian_media_and_rf.media_tensors.media_spend[
+                ..., -15:, :
+            ],
+            reach=self.meridian_media_and_rf.rf_tensors.reach[..., -15:, :],
+            frequency=self.meridian_media_and_rf.rf_tensors.frequency[
+                ..., -15:, :
+            ],
+            rf_spend=self.meridian_media_and_rf.rf_tensors.rf_spend[
+                ..., -15:, :
+            ],
+            revenue_per_kpi=self.meridian_media_and_rf.revenue_per_kpi[
+                ..., -15:
+            ],
+        )
+    )
+    self.assertAllClose(
+        roi,
+        tf.convert_to_tensor(test_utils.ROI_NEW_TIMES_DATA),
+        rtol=1e-3,
+        atol=1e-3,
+    )
 
   @parameterized.product(
       use_posterior=[False, True],